Added Save Features
Added ability to save data from Measure and PlotProfile functions as txt
files. All functions save images or data to an external folder named
ImageJ on the android device instead of a random photo folder, each
function saves to its own individual folder inside of the ImageJ
folder(example ImageJ/PlotProfile). When operations that save an image
or the "Save Data" button is pressed, the location to find these objects
is listed at the top of the application.